[systemcloud] Include validation within monitor and stop actions
Pacemaker seems not to bother explicitly validating the parameters
before attempting to start a resource.

Include a validation step within the monitor action so that validation
failures are caught during the initial probe.  Include the same
validation step within the stop action so that validation errors from
the initial probe are not overwritten by exceptions arising (due to
the invalid state) on the subsequent stop.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>