[rabbitmq] Rewrite RabbitMQ agent to actively manage cluster membership
The cluster_nodes configuration parameter is used only when first
starting a completely uninitialised node.  For nodes that have already
been started, the parameter is silently ignored by RabbitMQ.

Switch to a model similar to that for Galera, in which we first
identify the best bootstrap node (based on finding the last node to
shut down, or as close as possible if the cluster did not shut down
cleanly), then bringing up all remaining nodes as members (re)joined
to the bootstrap node.

We do not yet handle synchronisation of queues before demotion.
Demoting a node which is currently the master for an unsynchronised
queue may therefore leave the queue in a hung state.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>