[rabbitmq] Fix forgetting of removed cluster nodes
Removed nodes may be forgotten only after starting the bootstrap node,
and must be specified as a rabbit name list (i.e. a set of
"rabbit@xxx" strings) rather than a peer list.

Modify the join() and rejoin() methods to accept rabbit names rather
than peers, for consistency with forget().

With this change, it is possible to administratively remove a
permanently unrecoverable node via

  crm_node -R <node> -f

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>