[galera] Monitor wsrep_local_state in order to detect failed masters
Use SHOW STATUS to check wsrep_local_state as part of checking whether
or not the master is running.

This requires the ability to execute SQL commands at runtime.  To
avoid the need to store plaintext passwords, we install auth_socket.so
and grant the "mysql" user permission to connect locally,
authenticated via SCM_CREDS.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>