[systemcloud] Skip bootstrap recalculation when masters exist
A node may be left in the slave state simply because there are already
sufficient active masters to satisfy the master-max quota.  Such a
node should not attempt to trigger promotion of a bootstrap node,
since bootstrapping has already completed.

Fix by checking for an empty (or zero) master score before attempting
bootstrapping from the monitor action.  Note that we cannot check
instead for an empty list of current masters, since the monitor action
does not receive the notification environment variables.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>