[systemcloud] Prevent concurrent bootstrapping of multiple nodes
A brief network interruption can lead to a situation in which all
nodes have been demoted but none have yet been stopped.  Since the
master scores are preserved, this then leads to multiple nodes being
promoted simultaneously.

Fix by explicitly enforcing the restriction that only the bootstrap
node may be promoted when there are no existing masters.  We enforce
this as a temporary failure, to avoid any potential race conditions
around modifying the master score.

There is an uncomfortable amount of asymmetry surrounding the handling
of the bootstrapping process, but this essentially arises due to the
underlying inconsistencies in how information is provided to the
resource agent.  For example: the monitor action needs to be able to
trigger bootstrapping in the absence of any information about which
nodes are currently running as masters, since notifications are not
present for the monitor action.

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>