[systemcloud] Prevent services from starting outside our control
Create a flag file in /run and use an ExecStartPre check to ensure
that the flag file exists and is sufficiently recent before allowing
any controlled service to start.  This is designed to prevent
accidental service startup (e.g. due to a service being incorrectly
enabled to start at boot time).

Signed-off-by: Michael Brown <mbrown@fensystems.co.uk>